+++ src/reducers/handlers/reportHandlers.ts
@@ -1,15 +1,15 @@
 import { reportTypes, ERROR_UNKNOWN } from "../../types"
 
-export const reportsHandlers = {}
+export const reportHandlers = {}
 
-reportsHandlers[reportTypes.FETCH_ALL_SUCCESS] = function (state, action) {
+reportHandlers[reportTypes.FETCH_ALL_SUCCESS] = function (state, action) {
 	return {
 		reports: action.payload?._embedded?.reportInformationList,
 		error: "",
 	}
 }
 
-reportsHandlers[reportTypes.FETCH_ALL_FAILURE] = function (state, action) {
+reportHandlers[reportTypes.FETCH_ALL_FAILURE] = function (state, action) {
 	switch (action.payload.error) {
 		default:
 			return {
