+++ src/reducers/reportReducer.ts
@@ -1,0 +1,9 @@
+import { reducerFactory } from "./reducerFactory"
+import { reportHandlers } from "./handlers/reportHandlers"
+
+export const initialState = {
+	reports: null,
+	error: "",
+}
+
+export default reducerFactory(initialState, reportHandlers)
