--- src/reducers/reportsReducer.ts
@@ -1,9 +1,0 @@
-import { reducerFactory } from "./reducerFactory"
-import { reportsHandlers } from "./handlers/reportsHandlers"
-
-export const initialState = {
-	reports: null,
-	error: "",
-}
-
-export default reducerFactory(initialState, reportsHandlers)
